Refactor test modules
